Extract LegacyStripePaymentMethodsProvider
