+++ config/services/twig_extensions.php
@@ -3,11 +3,9 @@
 declare(strict_types=1);
 
 use FluxSE\SyliusStripePlugin\Stripe\SecretKey\LegacyKeyDetectorInterface;
+use FluxSE\SyliusStripePlugin\Stripe\SecretKey\LegacyStripePaymentMethodsProviderInterface;
 use FluxSE\SyliusStripePlugin\Twig\Extension\LegacyStripeKeyExtension;
-use Sylius\Bundle\PaymentBundle\Provider\GatewayFactoryNameProviderInterface;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
-
-use function Symfony\Component\DependencyInjection\Loader\Configurator\param;
 use function Symfony\Component\DependencyInjection\Loader\Configurator\service;
 
 return static function (ContainerConfigurator $container): void {
@@ -16,9 +14,7 @@ return static function (ContainerConfigurator $container): void {
     $services->set('flux_se.sylius_stripe.twig.extension.legacy_stripe_key', LegacyStripeKeyExtension::class)
         ->args([
             service(LegacyKeyDetectorInterface::class),
-            service('sylius.repository.payment_method'),
-            service(GatewayFactoryNameProviderInterface::class),
-            param('flux_se.sylius_stripe.factories'),
+            service(LegacyStripePaymentMethodsProviderInterface::class),
         ])
         ->tag('twig.extension');
 };
